Issue #270 - TaskHeader, TaskTabs, Task
Created TaskHeader component;
Changes in TaskTabs component;
Changes in Task component;
Removed TaskTabs container;
