--- frontend/src/containers/task-tabs.js
@@ -1,10 +1,0 @@
-import { connect } from 'react-redux'
-import TaskTabs from '../components/task/task-tabs'
-
-const mapStateToProps = (state, ownProps) => {
-  return {
-    user: state.loggedIn.user
-  }
-}
-
-export default connect(mapStateToProps)(TaskTabs)
